+++ src/emit/structure.rs
@@ -383,6 +383,9 @@ fn emit_part_def_body_element(
         PartDefBodyElement::ConstraintUsage(c) => {
             super::view::emit_constraint_usage(w, path, &c.value)
         }
+        PartDefBodyElement::RequireConstraint(c) => {
+            super::requirement::emit_require_constraint(w, path, &c.value)
+        }
         PartDefBodyElement::AssertConstraint(a) => {
             super::view::emit_assert_constraint(w, path, &a.value)
         }